Added the protobuf to define the state of overlay networks.
The JSON schema allows each agent to  maintain state about the
overlays networks that it is hosting. It will also be used by the
Master to communication configuration parameters for each overlay
instance to Agents. Configuration parameters would be: overlay subnet,
subnet assigned to an Agent for a given overlay, and the backend used to
setup the transport for the overlay (for example VxLAN).