Fixed CORE-397. (#7)
In `_configureNetwork`, once the Agent detects an error while
configuring the network, it should set the state of the configure
network as ERROR and return, rather than fall through.

Also, in `configureNetwork` replaced all returns with `Failure` with
calls to `_configureNetwork`. The reason being that a `Failure` is
defined as the network not being added to the database. However, once
the network is added to the database any failure in network
configuration should be captured by setting the state of the network
to "ERROR".