Introduced protobuf to store Agent configuration and overlay options.
The `AgentConfig` protobuf is used to store Agent configuration, as
specified by the operator, in the JSON configuration of the module. The
`AgentNetworkConfig` is used to inform the Agent (so that it can
inform the Master) if it's acting as a host for containers attached to
the overlay or if it is acting as gateway for DC/OS components to
talk to containers on the overlay.