Fixed re-registration of Agent when connectivity with Master is lost.
Currently, the agent `state` was not reverting to `REGISTERING` and
the registration was not being triggered when the agent looses
connectivity with the Master. Hence, if the Master re-starts the agent
module was not re-registering.

Set the `state` of the agent module to `REGISTERING` and triggered the
reliable registration from the agent so that it can re-register. In
case of change in leadership of the master, the reliable registration
would fail, and the `MasterDetector` would detect a new master
allowing the agent module to register with the new master.