Re-factored overlay configuration and introduced IPMASQ per overlay.
Currently we have an iptables masquerade rule for the entire overlay.
However, we want to rule to be applied only to the overlay traffic,
hence have to set the rule to filter based on source IP address, which
will be the overlay network subnet. We need to install a rule for each
overlay network.

Introducing the iptables masquerade rule complicates the Agents
execution logic and hence we have done the following re-factoring to
simplify the Agent:
* Removed `required` clause for STATUS:
Overlay networks by default should be in INVALID state whenever their
state is initialized. Hence, made the field optional and set a default
value for this field.

* Re-factored the overlay updates to capture any errors:
Since we weren't analyzing the `Futures` returned by the `await`, if
an error occurred when executing any of the `ipset` or `iptable`
commands, during the configuration of Mesos or Docker networks
currently we would simply not know about it. Capturing the `Futures`
returned by the await and logging any errors before sending updates
to the Master.

* Re-factored the `OverlayAgentManager` to use scripts:
We need to execute multiple ipset and iptables commands to setup IP
masquerade for the containers. Executing them one by one becomes
painful since this results in deep nesting of the `Futures`. Instead
we opted to combine these calls into a single script and execute the
script.