Updated the `State` of each overlay instance on the Agent.
We weren't updating the `State` of an overlay instance on the Agent,
when the Agent was acknowledging the configuration that the Master had
downloaded to the Agent. This was resulting in the
`overlay-master/state` endpoint not displaying the status of the
overlay instance on an Agent.

We now update the `State` of an overlay instance when we get an
acknowledgement from the Agent. We however do not store this state in
the replicated log, since the state is dependent on Agent registration
and is dynamic. Even if the `State` is written into the replicated
log, we clear the state during recovery.