Fixed the inconsistency between `AgentInfo` exposed by Master and Agent.
The `overlay-master/state` endpoint was exposing only the IP of the
Agent in the `ip` field of `AgentInfo`. However, the
`overlay-agent/overlay` endpoint was exposing the IP and port of the
Agent in the `ip` field of `AgentInfo`.

Made this behavior consistent across the two endpoints by exposing
only the IP of the Agent in the `ip` field of `AgentInfo`.