Modified the `MasterManagerProcess` to support batch writes to log.
Currently, for each agent registration we issue a write to the overlay
replicated log. This can degrade performance of the replicated log
when the number of agents is large. We have therefore introduced batch
writes to the overlay replicated log.

When an agent registers, if there are no pending writes to the
replicated log, the `Operation` used to mutate the `networkState` with
the new `AgentInfo` associated with the agent will be added to the
`operations` queue. The `MasterManagerProcess` will then update the
replicated log with a copy of the updated `networkState`. Once the
replicated log has been successfully updated, the `State` will be
reflected in the `networkState` and a configuration update is sent to
every agent on the list.

As part of these changes have also modified the variables used to
store the state information during "recovery" of the replicated log.