Added metrics prefix to the overlay replicated log.
Currently we don't setup a prefix for the overlay replicated log,
resulting in the overlay replicated log showing up as 'log/recover' in
the metrics end-point. This becomes very cryptic when multiple
replicated logs are in play. Therefore given a meaningful prefix to
the overlay replicated log prefix to make it easily identifiable.