Modified agent to better handle failure dependencies.
Instead of failing hard when overlay configuration fails, the agent
will now complete the registration with the Master, and allow the
agent to proceed. The state of the overlay will be reflected on the
Master end-point and it will be the responsibility of the operator to
fix the error and manually restart the agent in order to re-configure
the overlay.