Modified overlay tests to use `agent::ManagerProcess`.
Currently the tests were instantiating an overlay agent module for
tests. The side-affect of this was that the module creation blocks on
the successful registration of the agent module with the Master before
returning. This has the unfortunate side-affect of making a test hang
if for some reason the agent was unable to register with the Master.

Instead of instantiating an agent overlay module we are therefore
directly instantiating an agent overlay process, so that we can access
the futures within the agent overlay process and not block on agent
registration.