Modified the re-registration logic to allow change to agent config.
Added an `UpdateAgent` `operation` to the overlay replicated log to
update `AgentInfo` for agents whose overlay configuration might need
to be changed due to a change in the agent's network config. The
change in agent's network config could be the removal or addition of a
`docker_bridge` or a `mesos_bridge` and the addition or removal of an
`allocated_subnet`.