Added a uni-test to check agent `network_config` change.
Added a test to check that the Master is honoring any `network_config`
change on the agent. For e.g., if the agent has disabled its Mesos or
Docker network the Master should be honoring it by disabling any Mesos
or Docker bridges in its existing overlay configuration for the Agent.