Modified Mesos and Docker configuration in the Agent.
Instead of relying on the information sent by the Master, we create
Mesos and Docker networks based on the configuration given by the
operator.

With this change the semantics of Master Agent interaction change to
the overlay configuration being sent to the agent being treated as a
request that the agent will try to comply to as long as the operator
allows the agent to apply the configuration requested by the Master.