Removed dependency of Agent startup on overlay configuration.
Currently, the agent overlay module was blocking the agent from
starting till the overlay configuration on the agent was complete.
This was primarily being done due to the limitation of the
`network/cni` isolator to dynamically ADD/DELETE/MODIFY CNI
configuration without an agent restart. This limitation on the
`network/cni` isolator has been removed with MESOS-6567. Therefore,the
agent overlay module no longer needs to block the agent and force it
to wait for any CNI configurations of the overlay network to be
available.