Modified behavior of agent module on failed overlay configuration.
Currently, the overlay module would exit if it fails to configure all
the overlays after `maxConfigAttempts`. This was primarily done to
force an agent restart in case of overlay configuration failure since
the agent was blocked on the overlay module bootup.

Since we have removed the dependency of the agent startup on the
overlay module bootup, we no longer need to force the agent restart on
an overlay configuration failure. The overlay module can remain in an
erroneous state (allowing the agent to perform other functions), and
the agent can be restarted by the operator, manually, once the
configuration errors for the overlay have been rectified.

Have also fixed a bug in the book keeping of `configAttempts`. We
weren't resetting the `configAttempts` when we detect a new Master or
when the older Master exists. This was forcing the overlay module to
FAIL, even when valid overlay configuration existed, when multiple
Master switchover takes place.