Fixed a CHECK failure when using Docker Containerizer.
When you pass the Docker Containerizer a Task/Executor ID with a
colon in the ID, the sandbox path will be changed from the
canonical path to a symlink.  This symlink is missing most
of the information of the canonical path.

The journald logger was relying on the sandbox path to derive
the Agent/Container IDs.  This commit changes the logic to only
add Agent/Container ID labels when the fields can be derived,
instead of CHECK-ing that the fields are present.