Added configurable limit to the size of labels passed in logger.
This adds a rather primitive limit to the number or size of
labels passed from tasks to the container logger companion
binary.  Once the JSON-ified size of labels surpasses a
configurable size, the module simply cuts off any more
arbitrary labels before proceeding with the logger module's
business as usual.