Fixed segfault in flag creation during tests.
This adds the `-DUSE_SSL_SOCKET` definition to the build at
configuration time, in line with how the Mesos Autotools build adds
this definition for SSL-enabled builds.  Because DC/OS always compiles
with Libevent and SSL enabled, we will always need this definition.

The segfault occurs when calling the rather innocent looking constructor
for the Mesos Agent flags: `mesos::internal::slave::Flags::Flags()`.
Under the hood, `Flags` is a super-class of the stout `FlagsBase`
template class.  Each time we go to add a new field to the `Flags`,
the `FlagsBase` attempts to up-cast (via `dynamic_cast`) from
`FlagsBase` to the templatized `Flags` type.  `dynamic_cast`
will return `nullptr` if the cast fails, hence the segfaults.

The reason why the up-cast was failing is because the agent flags
include an `#ifdef USE_SSL_SOCKET`, which means that the `Flags`
with and without `USE_SSL_SOCKET` are actually different classes.
This meant the module tests were attempting to construct non-SSL
agent flags, while linking to SSL agent flags.