Add checkAgentResetConfigurationAttempts test.
Currently, if there are multiple configuraton attempt failures for the
overlay module running on the agent, when registered with the overlay
module in the Master it leads to the overlay modules actor just giving
up on the registration.

We need to test that the number of configuration attemtps are reset when
master failover happens, i.e., if a master failover happens and number
of config attempts is less than the number of attempts supported by the
overlay module the number of config attempts should be reset and the
overlay module should continue attempting to register.

JIRA: https://jira.mesosphere.com/browse/DCOS_OSS-728