Fixed flaky JournaldLoggerTest.ROOT_LogToJournald test.
Due to the lack of synchronization between container logger cleanup and
task termination, a write to the task's stdout or stderr log might occur
after a terminal status update is sent.

To adjust a test's expectation without introducing synchronization, a
busy loop with a timeout is added to check for the expected log output.