Added a note about launcher sealing and libtool wrapper build flags.
The JournaldContainerLogger module will depend on a function built
in Mesos, only when --enable-launcher-sealing is turned on.
This build flag also requires --disable-libtool-wrappers (otherwise,
many other things will break).  There is currently no good way to
enforce these requirements besides documenting them.

If these flags are not enabled in the Mesos build, there will be
a link error in the modules, and tests will fail.