Made logging module use sandbox logging for DEBUG containers.
Perf traces from DC/OS agents running a considerable number of command
health checks showed that the agent spends a considerable amount of time
forking out processes to pipe the output of `DEBUG` containers to
journald.

`DEBUG` containers are mostly used by the default executor to run checks
and by operators using `dcos task exec` to run commands in interactive
sessions. The default executor already logs the output of the checks it
performs, and operators most probably don't expect the interactive
sessions logs to be sent to journald, so we decided to use sandbox
logging for `DEBUG` containers.