Windows: Fixed reading of STDIN with the native event loop.
I/O completion ports (IOCP)  are the backbone of Mesos' implementation
of a native event loop on Windows.  It happens that IOCP cannot be
used with non-disk files, like the standard input/output handles.

Instead, we simply need to make blocking reads on STDIN.  Considering
what the logger binary needs to do, this is not a big demerit.